feat: order-inventory integration via WebClient (#2)
* feat: add order-inventory integration via WebClient

  - WebClient calls inventory-service reserve/release endpoints on placeOrder
  - Rollback releases all previously reserved items if any reservation fails
  - Add InventoryUnavailableException (409) with GlobalExceptionHandler mapping
  - Add StockReservationRequest DTO and WebClientConfig bean
  - Change productId from String to Long in OrderItem + migration V2
  - Change stock from BigDecimal to Integer in inventory-service + migration V2
  - Add WebClient mock chain to OrderServiceTest with failure/rollback tests

* test: add WebClient unit and WireMock integration tests

+++ inventory-service/src/main/java/com/example/inventory/repository/inventory/dto/ProductResponse.java
@@ -2,14 +2,12 @@ package com.example.inventory.repository.inventory.dto;
 
 import com.example.inventory.repository.inventory.model.Product;
 
-import java.math.BigDecimal;
 import java.time.Instant;
-import java.util.List;
 
 public record ProductResponse(
         Long id,
         String name,
-        BigDecimal stock,
+        Integer stock,
         Instant createdAt,
         Instant updatedAt
 ) {
