+++ inventory-service/src/main/java/com/example/inventory/repository/inventory/dto/UpdateStockRequest.java
@@ -2,9 +2,7 @@ package com.example.inventory.repository.inventory.dto;
 
 import jakarta.validation.constraints.NotNull;
 
-import java.math.BigDecimal;
-
 public record UpdateStockRequest(
         @NotNull(message = "El stock a reservar/liberar es requerido")
-        BigDecimal stock
+        Integer stock
 ) {}
